+++ controller/hybridgateway/watch/watch.go
@@ -74,6 +74,18 @@ func Watches(obj client.Object, cl client.Client) []Watcher {
 				MapHTTPRouteForKongUpstreamPolicy(cl),
 				&configurationv1beta1.KongUpstreamPolicy{},
 			},
+			{
+				MapHTTPRouteForClientCertSecret(cl),
+				&corev1.Secret{},
+			},
+			{
+				MapRouteForKongResource[*configurationv1alpha1.KongCertificate](kindHTTPRoute),
+				&configurationv1alpha1.KongCertificate{},
+			},
+			{
+				MapRouteForKongResource[*configurationv1alpha1.KongReferenceGrant](kindHTTPRoute),
+				&configurationv1alpha1.KongReferenceGrant{},
+			},
 		}
 	case *gwtypes.TLSRoute:
 		return []Watcher{
@@ -112,6 +124,18 @@ func Watches(obj client.Object, cl client.Client) []Watcher {
 			{
 				MapTLSRouteForReferenceGrant(cl),
 				&gwtypes.ReferenceGrant{},
+			},
+			{
+				MapTLSRouteForClientCertSecret(cl),
+				&corev1.Secret{},
+			},
+			{
+				MapRouteForKongResource[*configurationv1alpha1.KongCertificate](kindTLSRoute),
+				&configurationv1alpha1.KongCertificate{},
+			},
+			{
+				MapRouteForKongResource[*configurationv1alpha1.KongReferenceGrant](kindTLSRoute),
+				&configurationv1alpha1.KongReferenceGrant{},
 			},
 		}
 	case *gwtypes.Gateway: